pyproject.toml
[tool.poetry]
name = "rag-project"
version = "0.1.0"
description = "OpenAI와 Ollama(로컬 LLM) 모두 지원하는 한국어 최적화 RAG 시스템"
authors = ["blacktop39 <blacktop39@users.noreply.github.com>"]
homepage = "https://github.com/blacktop39/rag"
repository = "https://github.com/blacktop39/rag"
keywords = ["rag", "llm", "ollama", "openai", "chatbot", "korean", "medical"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.9"
openai = "^1.35.13"
langchain = "^0.2.7"
langchain-community = "^0.2.7"
langchain-openai = "^0.1.14"
chromadb = "^0.5.3"
sentence-transformers = "^3.0.1"
pypdf2 = "^3.0.1"
python-dotenv = "^1.0.1"
fastapi = "^0.112.0"
uvicorn = "^0.30.1"
pydantic = "^2.8.2"
langchain-ollama = "^0.1.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.0.0"
pytest-asyncio = "^0.23.0"
black = "^24.0.0"
isort = "^5.13.0"
flake8 = "^7.0.0"
mypy = "^1.8.0"

[tool.poetry.scripts]
rag-example = "example_usage:main"
medical-chatbot = "src.medical_chatbot:main"
medical-demo = "medical_chatbot_demo:run_demo"
local-llm-demo = "local_llm_demo:main"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true
strict_equality = true

[[tool.mypy.overrides]]
module = [
    "chromadb.*",
    "sentence_transformers.*",
    "langchain.*",
    "langchain_community.*",
    "langchain_openai.*"
]
ignore_missing_imports = true